Move changelogs-related code inside features/changelogs directory
